// data/github/current/ui/packages/query-builder/providers/__generated__/orgFilterProviderQuery.graphql.ts
export type orgFilterProviderQuery$variables = {
  count: number;
};
export type orgFilterProviderQuery$data = {
  readonly viewer: {
    readonly organizations: {
      readonly edges: ReadonlyArray<{
        readonly node: {
          readonly avatarUrl: string;
          readonly login: string;
          readonly name: string | null | undefined;
        } | null | undefined;
      } | null | undefined> | null | undefined;
    };
  };
};
export type orgFilterProviderQuery = {
  response: orgFilterProviderQuery$data;
  variables: orgFilterProviderQuery$variables;
};